requirements.txt
flask==2.0.1
werkzeug==2.0.3
pytesseract==0.3.8
Pillow==9.0.0
flask-cors==3.0.10 